Change behaviour of deleting users in regards to archived events + Fix various bugs + More tests

+++ server/db_scripts/delete_event.sql
@@ -18,6 +18,10 @@ begin
     raise exception 'Only event admins can delete event' using errcode = 'P0085';
   end if;
 
+  if exists (select 1 from "event" e where e.id = ip_event_id and e.active = false) then
+    raise exception 'Archived events cannot be deleted' using errcode = 'P0119';
+  end if;
+
   delete from "event" e where e.id = ip_event_id;
 
 end;
